requirements.txt
Flask~=2.2
gunicorn~=20.1
Werkzeug~=2.2
requests~=2.28
flask-sqlalchemy~=3.1
psycopg2-binary~=2.9
python-dotenv~=1.0
bcrypt~=4.1